feat: statusicon refactor
