+++ apps/cowswap-frontend/src/common/hooks/useWalletStatusIcon.ts
@@ -1,11 +1,11 @@
 import { useWalletDetails } from '@cowprotocol/wallet'
 import { useWeb3React } from '@web3-react/core'
 
-import { getStatusIcon } from 'modules/account/containers/AccountDetails'
+import { StatusIcon } from 'modules/account/containers/AccountDetails/StatusIcon'
 
 export function useWalletStatusIcon(): JSX.Element | null {
   const walletDetails = useWalletDetails()
   const { connector } = useWeb3React()
 
-  return getStatusIcon(connector, walletDetails, 56)
+  return <StatusIcon connector={connector} walletDetails={walletDetails} size={56} />
 }
