+++ apps/cowswap-frontend/src/legacy/components/TransactionConfirmationModal/LegacyConfirmationPendingContent.tsx
@@ -22,7 +22,7 @@ import SVG from 'react-inlinesvg'
 
 import { MediumAndUp, useMediaQuery } from 'legacy/hooks/useMediaQuery'
 
-import { getStatusIcon } from 'modules/account/containers/AccountDetails'
+import { StatusIcon } from 'modules/account/containers/AccountDetails/StatusIcon'
 
 import {
   ApproveComparison,
@@ -161,7 +161,9 @@ export function LegacyConfirmationPendingContent({
       <UpperSection>
         <CloseIconWrapper onClick={onDismiss} />
 
-        <WalletIcon>{getStatusIcon(connector, walletDetails, 56)}</WalletIcon>
+        <WalletIcon>
+          <StatusIcon connector={connector} walletDetails={walletDetails} size={56} />
+        </WalletIcon>
         <span>{pendingText}</span>
       </UpperSection>
