+++ apps/cowswap-frontend/src/modules/account/containers/AccountDetails/index.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import { Fragment, useMemo } from 'react'
 
 import { CHAIN_INFO } from '@cowprotocol/common-const'
@@ -9,24 +8,19 @@ import {
   getExplorerAddressLink,
   isMobile,
 } from '@cowprotocol/common-utils'
-import { MouseoverTooltip } from '@cowprotocol/ui'
 import { ExternalLink } from '@cowprotocol/ui'
 import {
   ConnectionType,
   useWalletInfo,
-  WalletDetails,
-  getConnectionIcon,
   getConnectionName,
   getIsCoinbaseWallet,
   getIsMetaMask,
-  Identicon,
   useWalletDetails,
   useIsWalletConnect,
   getWeb3ReactConnection,
   getIsHardWareWallet,
 } from '@cowprotocol/wallet'
 import { useWeb3React } from '@web3-react/core'
-import { Connector } from '@web3-react/types'
 
 import { Trans } from '@lingui/macro'
 
@@ -41,11 +35,11 @@ import Activity from 'modules/account/containers/Transaction'
 
 import { useIsProviderNetworkUnsupported } from 'common/hooks/useIsProviderNetworkUnsupported'
 
+import { StatusIcon } from './StatusIcon'
 import {
   AccountControl,
   AccountGroupingRow,
   AddressLink,
-  IconWrapper,
   InfoCard,
   LowerSection,
   NetworkCard,
@@ -78,66 +72,6 @@ export function renderActivities(activities: ActivityDescriptors[]) {
         return <Activity key={activity.id} activity={activity} />
       })}
     </TransactionListWrapper>
-  )
-}
-
-export function getStatusIcon(
-  connector: Connector,
-  walletDetails?: WalletDetails,
-  size?: number,
-  imageLoadError?: boolean,
-  setImageLoadError?: (error: boolean) => void
-) {
-  const connectionType = getWeb3ReactConnection(connector)
-
-  // Check for image load error
-  if (imageLoadError) {
-    const icon = getConnectionIcon(connectionType.type)
-
-    if (icon === 'Identicon') {
-      return <Identicon size={size} />
-    }
-
-    return (
-      <IconWrapper size={16}>
-        <img src={icon} alt={`${connectionType.type} logo`} />
-      </IconWrapper>
-    )
-  }
-
-  if (walletDetails && !walletDetails.isSupportedWallet) {
-    /* eslint-disable jsx-a11y/accessible-emoji */
-    return (
-      <MouseoverTooltip text="This wallet is not yet supported">
-        <IconWrapper role="img" aria-label="Warning sign. Wallet not supported">
-          ⚠️
-        </IconWrapper>
-      </MouseoverTooltip>
-    )
-    /* eslint-enable jsx-a11y/accessible-emoji */
-  }
-  if (walletDetails?.icon) {
-    return (
-      <IconWrapper size={16}>
-        <img
-          src={walletDetails.icon}
-          alt={`${walletDetails?.walletName || 'wallet'} logo`}
-          onError={() => setImageLoadError && setImageLoadError(true)}
-        />
-      </IconWrapper>
-    )
-  }
-
-  const icon = getConnectionIcon(connectionType.type)
-
-  if (icon === 'Identicon') {
-    return <Identicon size={size} />
-  }
-
-  return (
-    <IconWrapper size={16}>
-      <img src={icon} alt={`${connectionType.type} logo`} />
-    </IconWrapper>
   )
 }
 
@@ -181,9 +115,6 @@ export function AccountDetails({
 
   const unsupportedNetworksText = useUnsupportedNetworksText()
 
-  // State to track if the original image has failed to load
-  const [imageLoadError, setImageLoadError] = useState(false)
-
   function formatConnectorName() {
     const name = walletDetails?.walletName || getConnectionName(connection.type, getIsMetaMask())
     // In case the wallet is connected via WalletConnect and has wallet name set, add the suffix to be clear
@@ -222,7 +153,7 @@ export function AccountDetails({
                   }
                 }}
               >
-                {getStatusIcon(connector, walletDetails, 24, imageLoadError, setImageLoadError)}
+                <StatusIcon connector={connector} walletDetails={walletDetails} size={24} account={account} />
 
                 {(ENSName || account) && (
                   <WalletNameAddress>{ENSName ? ENSName : account && shortenAddress(account)}</WalletNameAddress>
